+++ tests/test_utils.py
@@ -120,7 +120,7 @@ def test_is_collection():
     [
         (dt.datetime(2013, 11, 10, 1, 23, 45), "Sun, 10 Nov 2013 01:23:45 -0000"),
         (
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, tzinfo=dt.timezone.utc),
             "Sun, 10 Nov 2013 01:23:45 +0000",
         ),
         (
@@ -138,11 +138,11 @@ def test_rfc_format(value, expected):
     [
         (dt.datetime(2013, 11, 10, 1, 23, 45), "2013-11-10T01:23:45"),
         (
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45, 123456)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, 123456, tzinfo=dt.timezone.utc),
             "2013-11-10T01:23:45.123456+00:00",
         ),
         (
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, tzinfo=dt.timezone.utc),
             "2013-11-10T01:23:45+00:00",
         ),
         (
@@ -161,7 +161,7 @@ def test_isoformat(value, expected):
         ("Sun, 10 Nov 2013 01:23:45 -0000", dt.datetime(2013, 11, 10, 1, 23, 45)),
         (
             "Sun, 10 Nov 2013 01:23:45 +0000",
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, tzinfo=dt.timezone.utc),
         ),
         (
             "Sun, 10 Nov 2013 01:23:45 -0600",
@@ -181,16 +181,16 @@ def test_from_rfc(value, expected):
         ("2013-11-10T01:23:45", dt.datetime(2013, 11, 10, 1, 23, 45)),
         (
             "2013-11-10T01:23:45+00:00",
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, tzinfo=dt.timezone.utc),
         ),
         (
             # Regression test for https://github.com/marshmallow-code/marshmallow/issues/1251
             "2013-11-10T01:23:45.123+00:00",
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45, 123000)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, 123000, tzinfo=dt.timezone.utc),
         ),
         (
             "2013-11-10T01:23:45.123456+00:00",
-            utils.UTC.localize(dt.datetime(2013, 11, 10, 1, 23, 45, 123456)),
+            dt.datetime(2013, 11, 10, 1, 23, 45, 123456, tzinfo=dt.timezone.utc),
         ),
         (
             "2013-11-10T01:23:45-06:00",